Add new struct "entryDetail" in "phonebook_opt.h"
The instruction of "cache-test" in Makefile is now added with

- L1-dcache-load-misses
- L1-dcache-prefetch-misses
- L1-dcache-store-misses

in order to measure the cache-misses of L1d cache.

"phonebook_opt.c" now has the same code as the "phonebook_orig.c"
to perform basic linked-list search and insert.

"phonebook_opt.h" is added with a struct named "entryDetail"
to store the data ohter than "lastName" within the struct "entry".
Therefore, the original struct "entry" now has three data:

- char lastName[MAX_LAST_NAME_SIZE];
- entryDetail* detail;
- struct __PHONE_BOOK_ENTRY *pNext;